(1248) Split up planShantytownModel
